Add files via upload
As a Data Scientist, This project aims to classify Customer Sentiment by predicting whether review is positive or negative based on customer reviews. Seaborn and matplotlib is used for visualization. The ultimate goal is to use pandas for clean, handle missing values and formatting  data for analysis. Counter Vectorizer is used to transform text documents into a matrix of token counts. Algorithm such as GaussianNB. Evaluation metrics such as accuracy and confusion matrix will measure performance.